refactor(mobile): extract ProjectScreen row components
Pull the notes-tree presentational pieces out of the 633-line ProjectScreen
into src/screens/project/: NoteRowItem, FolderRow, Segment, Empty (all
inline-styled, take a t prop), NoteFilterBar (own style slice), plus the
ListRow type + rowKey helper in list-rows.ts.

Host ProjectScreen.tsx drops to ~440 lines — state, load, long-press note
actions, move pickers, tab switch, and the row flattening + composition. No
behaviour or styling change; public export unchanged (only the two [id] route
files import it).

Verified: mobile tsc + eslint . clean; mobile vitest project (26) green.
