+++ mobile/changelogs/v0.1.3.md
@@ -67,3 +67,4 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - Split the Calendar view (previously a single ~930-line file) into focused pieces — the month/week grid math, the day cell, task chips, the unscheduled tray, and the day-detail row — with the grid/date logic now covered by unit tests. Mobile logic tests are also wired into the shared test runner for the first time (the calendar date math and the drag auto-scroll helpers now run in CI). No visible change; the calendar looks and behaves exactly the same.
 - Moved the on-device semantic-search index queries (note + task embeddings) out of the 1,600-line database-queries file into their own module, and factored the shared SQL fragments into one place. Purely internal reorganisation — every existing import keeps working via a re-export; no behaviour change.
 - Split the AI-settings form into smaller pieces (the provider/reasoning segmented buttons, the labelled input field, and the Private-Cloud-Compute usage bar now live in their own files, sharing one style module). Purely internal; the settings screen looks and behaves exactly the same.
+- Split the project detail screen — its note-row, folder-row, tab segment, filter bar and list-row helpers now live in their own files. Purely internal; the project's Overview / Notes / Board tabs look and behave exactly the same.
